Fix for proximity wakeup
The problem:
Panel is being set to "busy" state after executing double tap gesture.
If the phone is not woken up (i.e. PowerManager decided against it due to proximity sensor readings), panel is never reset to idle.
And because of that, all subsequent doubletap events are ignored.

The fix:
atomic_set panel to idle the panel after sending doubletap uevent but before "pressing" the power button.
This creates an unnecesary "free->busy->free->busy" transition, but it works.

Change-Id: I046a992769669355df74360b6ffbf5e3ffada1a6